requirements.txt
torch>=2.0.0
pyyaml>=6.0
numpy>=1.24.0
pytest>=7.0.0
black>=23.0
mypy>=1.0
tqdm>=4.65.0
# JAX backend (cuda12 works for both CPU and GPU)
jax[cuda12]>=0.4.0
flax>=0.7.0
optax>=0.1.7  # JAX optimizer library
